Add npm scripts to mangle markdown and yaml output into local static HTML pages that look somewhat like GitHub:
npm run all-for-local-html

This will render a readme.html in the root folder and html files in the docs folder for copying to some Web server.

It's important to note that the markdown expects references to the repo on GitHub; as such, the mangling necessairly changes references and links.

The "production" docs are still generated for GitHub using:

npm run all-for-github